moving queue implementation, handling error
