+++ README.md
@@ -2,8 +2,8 @@
 Websocket server to use with markdown editor in browser [here](https://github.com/iamkahvi/markdown_editor).
 
 ## Development
-`go run main.go` listens on port 8000
+`go run .` listens on port 8000
 
 ## Routes
-`/` - homepage  
+`/` - homepage
 `/write` - websocket
